pyproject.toml
[project]
name = "nps-automatico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.133.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langfuse>=3.14.5",
    "langgraph-cli[inmem]>=0.4.12",
    "langgraph-supervisor>=0.0.31",
    "msgpack>=1.1.2",
    "msgpack-types>=0.7.0",
    "pika>=1.3.2",
    "pip>=26.0.1",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "uuid>=1.30",
    "uvicorn>=0.41.0",
]
